Assembler: rewrite assembler
This initially started as a rewrite to have a classical
token -> parse -> emit pipeline, but after some consideration I decided
to keep it as a single run and combine all those steps into one.
This has cannibalised some of the old code but follows a new design I
consider to be better and more extensible.

Error reporting to be added.