VM: Add a memory mapped file backed binary manager
This commit also makes load_from use this new binary manager when our
file is too big for us to put it in RAM.
However, we no longer ever use the ifstream backed binary manager.
Though this commit doesn't remove it. Will keeping it prove to be a
maintanence burden? Who knows.

This commit does not implement error handling. That is to come.

Check out #2, the issue refering to what we are doing.